Transfers render board and tasks in module board
